requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
anthropic>=0.3.0
python-dotenv>=1.1.0
requests>=2.31.0
gitpython>=3.1.40
pathspec>=0.11.2
chardet>=5.0.0
langdetect>=1.0.9
tiktoken>=0.5.0
markdown>=3.4.0
markdown-extensions>=0.1.3